✨Add: Implement DDL and DML execution utilities with metadata support and create QueryResultBox component for displaying query results
