+++ frontend/apps/app/components/PGlitePage/index.ts
@@ -1,2 +1,3 @@
 export * from './PGlitePage'
 export * from './PGlitePlayground'
+export * from './QueryResultBox'
